requirements.txt
python-telegram-bot==20.8
requests
flask
gunicorn
python-dotenv
googlesearch-python